+++ sieve/las-types.h
@@ -86,7 +86,7 @@ struct sieve_info_s {
     // description of the q-lattice
     mpz_t q;
     mpz_t rho;
-    int32_t a0, b0, a1, b1;
+    int64_t a0, b0, a1, b1;
 
     // parameters for bucket sieving
     unsigned int td_thresh;
